JJH_Commit

+++ Movie_Project_spring/src/main/java/kr/ac/kopo/movie_project/dao/ServiceCenterDao.java
@@ -17,7 +17,7 @@ public interface ServiceCenterDao {
 
 	Board item(int boardId, int articleId);
 
-	void delete(int boardId, int articleId);
+	void delete(Long boardId, Long articleId);
